\name{conformity.Check}
\alias{conformity.Check}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Checks whether the dimension of marginal probability matrix matches the dimension of correlation matrix.
}
\description{
Checks whether the dimension of marginal probability matrix matches the dimension of correlation matrix.
}
\usage{
conformity.Check(ordPmat, CorrMat)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{ordPmat}{
Input matrix of ordinal marginal probabilities
}
  \item{CorrMat}{
Correlation matrix of the multivariate ordinal variable.
}
}